calc server

+++ resources/item.py
@@ -1,7 +1,8 @@
-import uuid
-from flask import request
+#import uuid
+#from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt  # x User autentication 
 from sqlalchemy.exc import SQLAlchemyError  # Exceptions used with SQLAlchemy.
 
 from db import db
@@ -17,16 +18,25 @@ blp = Blueprint("Items", __name__,  description="Operations on items")
 
 #----------------------------------------------------------------------------------------------
 
-@blp.route("/item/<string:item_id>")  #connect flask smorest with class Store bellow, so API runs the dinctions inside
+@blp.route("/item/<int:item_id>")  #connect flask smorest with class Store bellow, so API runs the dinctions inside
 class Item(MethodView): #class Store inherits from a methodView   
     """ITEM READ from db record --> db.session.add(item)""" 
+    
+    @jwt_required() # requests a token in header' Authetication -> Bearer'  with a token inside
     @blp.response(200, ItemSchema) # response deffinition  (includes dump_only and required from schema too into a response)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)   # Query.get but aborts with a 404 Not Found error instead of returning None. !! The primary key to query !!
         return item
     
-    """ITEM DELETE from db record --> """   
+    """ITEM DELETE from db record --> """  
+    @jwt_required() # requests a token in header' Authetication -> Bearer'  with a token inside
     def delete(self, item_id):
+        
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message=" Admin role is requested.")
+
+
         item = ItemModel.query.get_or_404(item_id) 
         db.session.delete(item)
         db.session.commit()
@@ -55,6 +65,7 @@ class Item(MethodView): #class Store inherits from a methodView
 @blp.route("/item")  
 class ItemList(MethodView): #class Store inherits from a methodView    
     """ITEM LIST ALL in db record --> """ 
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))  # after validation gives us a list sutible for return in function
     def get(self):
         #return("Hello word!")
@@ -63,6 +74,8 @@ class ItemList(MethodView): #class Store inherits from a methodView
         return ItemModel.query.all()
 
     """ITEM CREATE record --> db.session.add(item)"""
+    """@jwt_required(Fresh=True)  #A decorator to protect a Flask endpoint with JSON Web Tokens. Any route decorated with this will require a valid JWT """
+    @jwt_required()
     @blp.arguments(ItemSchema)   #here we DECORATED the metod with ItemSchema defined in schemas.py  for data valadation
     @blp.response(201, ItemSchema)
     def post(self, item_data): #++ item_data  ++  will be passsed after validation from ItemSchema
